+++ mrb_doc/core/cursor.rb
@@ -1,7 +1,0 @@
-# Checks if the cursor is in the window.
-# @return [Boolean]
-def cursor_on_screen?
-  # mrb_cursor_on_screen
-  # src/mruby_integration/core/cursor.cpp
-  true
-end